version 1.2
Add View, Create, Edit, Delete document pages

+++ DocuDash-API/Backend-API/src/main/java/doc/backendapi/service/DocumentService.java
@@ -1,13 +1,17 @@
 package doc.backendapi.service;
 
-import doc.backendapi.Exception.ResourceNotFoundException;
+import doc.backendapi.DTO.DocumentDto;
 import doc.backendapi.entities.Document;
 import doc.backendapi.repositories.DocumentRepository;
 import lombok.RequiredArgsConstructor;
+import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -16,26 +20,27 @@ public class DocumentService {
     @Autowired
     private final DocumentRepository documentRepository;
 
-    public List<Document> getAllDocument() {
-        return documentRepository.findAll();
+    @Autowired
+    private ModelMapper modelMapper;
+
+    public List<DocumentDto> getAllDocument() {
+        return documentRepository.findAll().stream().map((element) ->
+                modelMapper.map(element, DocumentDto.class)).collect(Collectors.toList());
     }
 
-    public Document saveDocument(Document document) {
-        return documentRepository.save(document);
+    public DocumentDto saveDocument(DocumentDto documentDto) {
+        Document document = modelMapper.map(documentDto, Document.class);
+        Document savedDocument = documentRepository.save(document);
+        return modelMapper.map(savedDocument, DocumentDto.class);
     }
 
-    public Document updateDocument(int id, Document document) {
-        Document existingDocument = documentRepository.findById(id).orElse(null);
+    public Document updateDocument(int id, DocumentDto document) {
+        Document existingDocument = documentRepository.findById(id).orElseThrow(() ->
+                new ResponseStatusException(HttpStatus.NOT_FOUND,
+                        "Document with id " + id + " not found. Please ensure the document id is correct and try again."));
 
-        if (existingDocument == null) {
-            // handle the case where the document with the provided ID does not exist
-            // you could throw an exception or return null
-            throw new ResourceNotFoundException("Document", "id", id);
-        }
         if (document.getTitle() != null)
             existingDocument.setTitle(document.getTitle());
-        if (document.getFilePath() != null)
-            existingDocument.setFilePath(document.getFilePath());
         if (document.getDateAdd() != null)
             existingDocument.setDateAdd(document.getDateAdd());
         if (document.getDateUpdate() != null)
@@ -50,16 +55,20 @@ public class DocumentService {
             existingDocument.setFromSource(document.getFromSource());
         if (document.getStatus() != null)
             existingDocument.setStatus(document.getStatus());
-        if (document.getUploadBy() != null)
-            existingDocument.setUploadBy(document.getUploadBy());
         return documentRepository.save(existingDocument);
     }
 
     public void deleteDocument(int id) {
+        if (!documentRepository.existsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND,
+                    "Document with id " + id + " not found. Please ensure the document id is correct and try again.");
+        }
         documentRepository.deleteById(id);
     }
 
     public Document getDocumentById(int id) {
-        return documentRepository.findById(id).orElse(null);
+        return documentRepository.findById(id)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND,
+                        "Document with id " + id + " not found. Please ensure the document id is correct and try again."));
     }
 }
